+++ src/routes/admin-route.js
@@ -1,14 +1,14 @@
 const router = require("express").Router();
-const { addProfessor } = require("../services/admin-service");
 const {
   addProfessorValidator,
 } = require("../utils/validators/admin-validator");
 
 const { studentValidator } = require("../utils/validators/student-validator");
-const { createStudent } = require("../services/user-service");
+const { createStudent, addProfessor } = require("../services/admin-service");
 const { verifyToken } = require("../middlewares/auth");
 
-router.post("/student", verifyToken, studentValidator, createStudent);
-router.post("/professor",addProfessorValidator, addProfessor);
+
+router.post("/student",verifyToken('admin'),studentValidator, createStudent);
+router.post("/professor", addProfessorValidator, addProfessor);
 
 module.exports = router;
